-compact_library: library RT as double, and the last score divergence goes with it
The compact path was giving 5,985 peptides against the ordinary path's 6,089 -- 424 lost
and 320 gained. I first read that as borderline churn, then ran the control: two ORDINARY
runs of the same configuration are bit-identical, 6,089 shared, 0 lost, 0 gained. The
pipeline is deterministic, so the churn was a real divergence.

Measured, comparing the two bundles field by field:

  precursor_mz                     0 of 423,079 differ
  var_ms2_massdev_score            0% differ
  var_ms2_intensity_score          0% differ
  library_rt                 422,353 of 423,079 differ, median 6.15e-09
  delta_rt                         median 1.66 s

PeptideRec::rt was a float where the ordinary reader keeps a double. 6e-9 in normalised
iRT is six orders of magnitude too small to move an RT by 1.66 s directly -- but the CiRT
calibration is an iterative fit, so a perturbation that size flips borderline anchors in
and out and the fitted curve moves. norm_rt, delta_rt, the RT score and, through the
pass-2 extraction windows, every chromatogram-shape score move with it.

That m/z and intensity agree EXACTLY is what makes this attributable: RT was the only
input left that differed.

Costs 8 B per peptide (record 48 -> 56 B, +54.5 MB over 7.15M, 0.08% of peak). My first
comment claimed the alignment padding absorbed it for free; measured, it does not.

+++ src/odia_library.h
@@ -219,8 +219,8 @@ public:
   {
     auto& r = peptides_[idx(p)];
     r.precursor_mz = mz;
-    r.rt = float(rt);
-    r.drift_time = float(drift);
+    r.rt = rt;
+    r.drift_time = drift;
   }
   double precursorMz(Peptide p) const { return peptides_[idx(p)].precursor_mz; }
   double rt(Peptide p) const { return peptides_[idx(p)].rt; }
@@ -327,8 +327,18 @@ private:
     SequenceStore::Span span;                              // sequence, as a slice of its protein
     std::uint32_t paired = ~std::uint32_t(0);   ///< a decoy's target, by peptide index
     double precursor_mz = 0.0;
-    float rt = std::numeric_limits<float>::quiet_NaN();    // library RT
-    float drift_time = -1.0f;
+    // DOUBLE, not float. It costs 8 B per peptide (the record goes 48 -> 56 B, +54.5 MB over
+    // 7.15M peptides, 0.08% of the run's peak) and it buys exact agreement with the ordinary
+    // reader, which is the whole point of this class.
+    //
+    // As a float these lost ~6e-9 of the library RT (measured: 422,353 of 423,079 precursors
+    // differed from the ordinary reader, median 6.15e-09). That is six orders of magnitude too
+    // small to matter directly, but the CiRT calibration is an iterative fit: a perturbation that
+    // size flips borderline anchors in and out, the fitted curve moves ~1.7 s, and every RT and
+    // chromatogram-shape score moves with it. The measured cost was 424 peptides lost and 320
+    // gained against a path that is otherwise bit-identical run to run.
+    double rt = std::numeric_limits<double>::quiet_NaN();  // library RT
+    double drift_time = -1.0;
     Protein parent = no_protein;
     SequenceStore::Id original_id = SequenceStore::npos;
     std::int16_t charge = 0;
